requirements.txt
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
scipy>=1.11
lightgbm>=4.0
joblib>=1.3
tqdm>=4.66
requests>=2.32
torch>=2.3
torchvision>=0.18
Pillow>=10.0
